requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
replicate==0.27.0
python-dotenv==1.0.1
